Refactored to use a view model
